% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondrian_nyc.R
\name{mondrian_nyc}
\alias{mondrian_nyc}
\title{mondrian_nyc}
\usage{
mondrian_nyc(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
df
}
\description{
mondrian_nyc
}
\examples{
df=data.frame(x=sample(18,5),y=sample(18,5))
mondrian_nyc(df)
}
