% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondrian_paris.R
\name{mondrian_paris}
\alias{mondrian_paris}
\title{mondrian_paris}
\usage{
mondrian_paris(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
graph
}
\description{
mondrian_paris
}
\examples{
df=data.frame(x=sample(6,3),y=sample(6,3))
mondrian_paris(df)

}
